Move searching test to search.spec.js file
