+++ test/tests.spec.js
@@ -3,3 +3,4 @@ require('./server/roles.spec.js');
 require('./server/users.spec.js');
 require('./server/authentication.spec.js');
 require('./server/documents.spec.js');
+require('./server/search.spec.js');
